recast numeric vars to floats

+++ app.py
@@ -28,9 +28,11 @@ df_final = df_final.drop(['ste_code','ste_name','ste_area_code','ste_type','ste_
 df_final = df_final.rename(columns={'period_begin':"Period",'property_type':"Type of Property",'median_sale_price':"Median Sale Price",
                                     'median_sale_price_yoy':"Median Sale Price YoY",'homes_sold':"Homes Sold",'state_code':"State"})
 
-df_final["Median Sale Price"] = df_final["Median Sale Price"].astype(int)
-df_final["Median Sale Price YoY"] = df_final["Median Sale Price YoY"].astype(int)
-df_final["Homes Sold"] = df_final["Homes Sold"].astype(int)
+#### MUST KEEP AS FLOATS, NO INT
+# df_final["Median Sale Price"] = df_final["Median Sale Price"].astype(int)   
+# df_final["Median Sale Price YoY"] = df_final["Median Sale Price YoY"].astype(int)
+# df_final["Homes Sold"] = df_final["Homes Sold"].astype(int)
+#### MUST KEEP AS FLOATS, NO INT
 df_final["Month"] = pd.to_datetime(df_final["Period"], format='%Y-%m-%d').dt.to_period('M')
 #df_final["Month"] = df_final["Month"].astype(str) ##(int) ####pd.to_datetime(df_final['Month'], format='%b %Y')
